抽离出公共的NetWorkManager
1:MultiBattle更名为NetWorkManager
2:原先MultiBattle针对太空站的逻辑放在了新脚本中：SpaceBattleManager
3：又出现了之前的第二次游戏出问题，尝试在Revc结果后就ClearBattle()，但还未测试
